% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDepend.R
\name{calcDepend}
\alias{calcDepend}
\title{Calculate species dependencies (i.e. conditional probabilities).}
\usage{
calcDepend(a, b)
}
\arguments{
\item{a}{The abundance of species "a".}

\item{b}{The abundance of species "b".}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Calculates the species dependencies for a given set of species abundances as
the conditional probability of two species.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (a, b) 
{
    return(length(a[(a + b) == 2])/sum(a))
  }

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
%% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}
