% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoCo.R
\name{CoCo}
\alias{CoCo}
\title{Generate a vector of co-occurrence counts from an occurrence matrix}
\usage{
CoCo(x, type = c("pos", "neg"))
}
\arguments{
\item{x}{Community matrix with species in columns.}

\item{type}{Type of co-occurrence count matrix to return.}
}
\value{
Returns either co-occurrences (pos) or checkers (neg) counts for all unique non-recursive species pairs%% ~Describe the value returned .
}
\description{
Models and plots a network model using the methods of Araujo et al. 2011.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
