\name{ComGenR-package}
\alias{ComGenR-package}
\alias{ComGenR}
\docType{package}
\title{
  A toolbox for conducting Community Genetics analyses in R.
}
\description{
  This package compiles functions for many analysis and modeling
  techniques used in Community Genetics, including: compositional 
  analyses, simulations and network modeling.
}
\details{
\tabular{ll}{
Package: \tab ComGenR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-06\cr
License: \tab What license is it under?\cr
}
}
\author{
Maintainer: Matthew K. Lau <mkl48@nau.edu>
}
\references{
Whitham et al. 2013
Araujo et al. 2011
Shuster et al. 2006
}
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
