% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coSym.R
\name{coSym}
\alias{coSym}
\title{Calculates the symmetry of species in a network.}
\usage{
coSym(x = "dependency network", zero.na = TRUE)
}
\arguments{
\item{x}{Dependency network.}

\item{zero.na}{LOGICAL: should zero values be removed?}
}
\value{
Generates a measure of co-occurrence matrix symmetry.
}
\description{
Calculates node level statistics describing the symmetry of species
interactions in the network using the methods of Araujo et al. 2011.
}
