% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscore.R
\name{cscore}
\alias{cscore}
\title{Calculates the C-Score of Stone and Roberts 1991.}
\usage{
cscore(x, cu.mat = FALSE)
}
\arguments{
\item{x}{Community matrix of species abundances or present absences with
species in columns.}

\item{cu.mat}{Logical: should the checkerboard unit matrix be returned?}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Calculates the C-Score of Stone and Roberts 1991.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
