% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpmCom.R
\name{gpmCom}
\alias{gpmCom}
\title{Community genotype-phenotype mapping function.}
\usage{
gpmCom(n = "number of species", het.values = c(5, 21), allelic.range = c(0, 3))
}
\arguments{
\item{n}{Number of species.}

\item{het.values}{Range of heterozygote values.}

\item{allelic.range}{Range of allelic values.}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Creates a set of phenotypes based on genotypic values for a community of
organisms (e.g. insects) based on the methods of Shuster et al. 2006. This
is a support function for the cgSim function.
}
\details{
This function provides a simple way to populate a set of phenotypic values
for a model community to be used in a simulation of community genetics
effects.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau %% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}
