% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpmTrees.R
\name{gpmTrees}
\alias{gpmTrees}
\title{Foundation tree species genotype-phenotype mapping function for cgSim.}
\usage{
gpmTrees(pheno = "phenotype of genotypes", reps = "replication")
}
\arguments{
\item{pheno}{A vector of phenotypic values for a set of genotypes.}

\item{reps}{Scalar to set the number of replicates for the genotypes.}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Generates a matrix of genotypic and phenotypic values using the methods of
Shuster et al. 2006 to be used in the cgSim function.
}
\details{
This function populates a matrix to be used as the foundation tree species
that is influencing a set of species in a community in the cgSim function.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
