% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.g.R
\name{mean.g}
\alias{mean.g}
\title{Calcualtes the mean for a set of genotype.}
\usage{
\method{mean}{g}(x = "community matrix", g = "grouping")
}
\arguments{
\item{x}{Community matrix with species in columns.}

\item{g}{Grouping (i.e. genotype) vector.}
}
\description{
Calcualtes the mean for a set of genotype.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau %% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}
