% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.R
\name{mgp}
\alias{mgp}
\title{Modified graph plotting for network models.}
\usage{
mgp(net = "species network", com = "community matrix", my.coord = "",
  loc = TRUE, v.scale = 3, v.min = 0.1, displaylabels = FALSE)
}
\arguments{
\item{net}{A network.}

\item{com}{The community matrix used to generate the network with species in
the same order as the network.}

\item{my.coord}{A set of point coordinates to coordinates the nodes in the
network.}

\item{loc}{Logical: should the location (i.e. coordinates) of the plotted
nodes be returned?}

\item{v.scale}{Value to scale the node (i.e. vertex) size.}

\item{v.min}{Minumum size for a node (i.e. vertex).}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
High level access to the gplot function in the sna package.
}
\details{
This is a high level function for plotting networks using the sna package
specifically tailored for community data based network models.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
