% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min.net.R
\name{min.net}
\alias{min.net}
\title{Minimized network function.}
\usage{
\method{min}{net}(net = "network", com = "community matrix")
}
\arguments{
\item{net}{A network model in matrix form.}

\item{com}{A community matrix with species in the same order as the network.}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Takes a network and community matrix reduces it to species that have
connections in the network.
}
\details{
This function is useful mainly for plotting.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
%% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}
