% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.prune.R
\name{null.prune}
\alias{null.prune}
\title{Pruning function for the Araujo et al. 2011 method.}
\usage{
null.prune(a, b, std = TRUE)
}
\arguments{
\item{a}{Species "a".}

\item{b}{Species "b".}

\item{std}{Logical: should the values be standardized?}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Defines whether or not a node should be included in a network using the
analytical methods of Araujo et al. 2011.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
