% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmTrees.R
\name{rmTrees}
\alias{rmTrees}
\title{Tree removal simulator.}
\usage{
rmTrees(
  x = "network",
  nits = 100,
  method = c("random", "degree", "type"),
  type = "grouping",
  return.nits = FALSE
)
}
\arguments{
\item{x}{A network in matrix representation.}

\item{nits}{Number of iterations to be used.}

\item{method}{Method to be used: 'random','degree','type'.}

\item{type}{Type of removel.}

\item{return.nits}{LOGICAL: should all iterations be returned.}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Removes trees from a matrix under three scenarios.
}
\details{
Removes trees from a simulation to simulate selection. Three
options for removal: random, degree and (geno- or pheno-)type
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
%% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}
