% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpp.R
\name{simSpp}
\alias{simSpp}
\title{Community simulator for running cgSim.}
\usage{
simSpp(n = "number of species", het.values = c(5, 21), allelic.range = c(0, 3))
}
\arguments{
\item{n}{Number of species.}

\item{het.values}{Range of heterozygote values.}

\item{allelic.range}{Range of allelic values.}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Generates a matrix of species allelic values.
}
\details{
This function provides a simple way to populate a set of phenotypic values
for a model community to be used in a simulation of community genetics
effects.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (n = "number of species", het.values = c(5, 21), allelic.range = c(0, 
    3)) 
{
    if (n == "number of species") {
        n <- 25
    }
    com <- matrix(NA, nrow = n, ncol = 2)
    com[, 1] <- runif(n, het.values[1], het.values[2])
    com[, 2] <- runif(n, allelic.range[1], allelic.range[2])
    com. <- com
    com.[, 1] <- (com[, 1] - 0.5 * com[, 2])/2
    com.[, 2] <- (com[, 1] + 0.5 * com[, 2])/2
    return(com.)
  }

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
%% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}
