% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrees.R
\name{simTrees}
\alias{simTrees}
\title{Tree simulator.}
\usage{
simTrees(tree.gpm = "tree genotype-phenotype map", VeT = 2)
}
\arguments{
\item{tree.gpm}{A genotype-phenotyp mapping matrix. See gpmTrees.}

\item{VeT}{Tree trait environmental variance.}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Generates a matrix of phenotypic values for a simulated forest of trees.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
%% ~~who you are~~
}
\keyword{~kwd1}
\keyword{~kwd2}
