% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch.plot.R
\name{ch.plot}
\alias{ch.plot}
\title{Cross-hair plotting functions for 2D NMDS plots.}
\usage{
ch.plot(
  x = "ordination matrix",
  g = "groupings",
  cex = 1,
  lwd = 1,
  mu.pch = 19,
  pt.col = 1,
  bar.col = 1,
  xlab = "X1",
  ylab = "X2"
)
}
\arguments{
\item{x}{Ordination matrix.}

\item{g}{Grouping (i.e. genotype) vector.}

\item{cex}{Value to scale centroids.}

\item{lwd}{Line width scaler.}

\item{mu.pch}{Centroid shape.}

\item{pt.col}{Centroid color.}

\item{xlab}{X-axis label: DEFAULT = X1.}

\item{ylab}{Y-axis label: DEFAULT = X2.}

\item{car.col}{Error bar color.}
}
\value{
Produces a 2-D plot of centroids with variance bars (i.e. "cross-hairs").
}
\description{
This is a high level plotting function that produces a "cross-hair" plot of
centroids and standard deviations for groups of ordinated points, as is
often desired with community genetics data.
}
\details{
Simple way to plot centroids of groups of ordinated points with standard
error values.
}
\author{
Matthew K. Lau
}
